+++ timelib.hpp
@@ -258,10 +258,10 @@ namespace tms
 		return (date);
 	}
 
-	short	difference_between_two_dates(tms::s_date date1, tms::s_date date2, bool include_last_day = false)
+	int	difference_between_two_dates(tms::s_date date1, tms::s_date date2, bool include_last_day = false)
 	{
-		short	difference = 0;
-	
+		int	difference = 0;
+
 		if (tms::is_date1_less_than_date2(date1, date2))
 		{
 			while (!tms::is_date1_equal_to_date2(date1, date2))
